Create a separate file for the fitting functions
